requirements.txt
Flask
gunicorn
eventlet
python-socketio